requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
aiogram==3.13.1
pydantic==2.9.2
python-dotenv==1.0.1